Moved element specific tests into the parsing package.
