+++ parse_test.go
@@ -1,21 +1,10 @@
 package main
 
 import (
-	"strconv"
 	"testing"
 
 	"github.com/billyedmoore/Slowdown/parser"
 )
-
-func TestParagraphSplitting(t *testing.T) {
-	lines := []string{"Line one", "Line two", "", "Second paragraph"}
-
-	root := parser.Parse(lines)
-
-	if len(root.GetChildren()) != 2 {
-		t.Fatalf("Should be 2 children instead found %d", len(root.GetChildren()))
-	}
-}
 
 func TestEmptyFileParsing(t *testing.T) {
 	lines := []string{}
@@ -43,63 +32,4 @@ func TestAllInlineBlocksBeingParsed(t *testing.T) {
 			t.Fatalf("Found an unparsed inline block after parsing")
 		}
 	}
-}
-
-func TestRawTextInParagraph(t *testing.T) {
-	lines := []string{"Line one", "", "Second paragraph"}
-	root := parser.Parse(lines)
-	// should be root.paragraph.rawText
-	firstContent := root.GetChildren()[0].GetChildren()[0].GetContent()
-
-	if firstContent != "Line one" {
-		t.Fatalf("Expected \"Line one\", instead got \"%v\"", firstContent)
-	}
-}
-
-func TestATXHeadingWithWhitespace(t *testing.T) {
-	lines := []string{"  #                    Line one        ", "", "Second paragraph"}
-	root := parser.Parse(lines)
-
-	firstBlock := root.GetChildren()[0]
-	firstContent := firstBlock.GetChildren()[0]
-
-	if firstContent.GetContent() != "Line one" {
-		t.Fatalf("Expected \"Line one\", instead got %v \"%v\"", firstBlock.GetNodeType(), firstContent.GetContent())
-	}
-}
-
-func TestATXHeadingWithMoreThan3Spaces(t *testing.T) {
-	lines := []string{"     #                    Line one        ", "", "Second paragraph"}
-	root := parser.Parse(lines)
-
-	firstContent := root.GetChildren()[0].GetChildren()[0]
-
-	if firstContent.GetNodeType() == "HEADING_BLOCK" {
-		t.Fatal("Expected a \"PARAGRAPH_BLOCK\" but got a \"HEADING_BLOCK\" instead")
-	}
-
-}
-
-func TestATXHeadingWithMoreThanSix(t *testing.T) {
-	lines := []string{"####### Line one", "", "Second paragraph"}
-	root := parser.Parse(lines)
-
-	firstContent := root.GetChildren()[0].GetChildren()[0]
-
-	if firstContent.GetNodeType() == "HEADING_BLOCK" {
-		t.Fatal("Expected a \"PARAGRAPH_BLOCK\" but got a \"HEADING_BLOCK\" instead")
-	}
-
-}
-
-func TestATXHeadingBreakingParagraph(t *testing.T) {
-	lines := []string{"Line one", "Line two", "# Heading", "Second paragraph"}
-	root := parser.Parse(lines)
-
-	firstContent := root.GetChildren()[0].GetChildren()[0].GetContent()
-
-	if firstContent != "Line one\nLine two" {
-		t.Fatalf("Expected a \"Line one\\nLine two\" but got a %v instead", strconv.Quote(firstContent))
-	}
-
 }
